docs(crewai): address code review feedback
- Rename provider.py to agent.py for clarity
- Fix skill matching logic inconsistency across test cases
- Add missing language specifiers to code blocks
- Add required bootstrap command to README
- Remove references to mock data - now uses real CrewAI output
- Update all model references to use gpt-4.1 consistently
- Use textwrap.dedent for cleaner multi-line string formatting
- Remove implementation detail references for evergreen documentation
